+++ sources/db_overall_plan.cpp
@@ -110,8 +110,10 @@ void DBOverallPlan::printBD() {
 }
 
 void DBOverallPlan::deleteDB(const std::string &name) {
-    const std::string DBName = path_ + config::separator + name;
-    fs::remove_all(DBName);
+    if (nameOpenDB_ != name){
+        const std::string DBName = path_ + config::separator + name;
+        fs::remove_all(DBName);
+    }
 }
 
 void DBOverallPlan::open(const std::string &name) {
